Feature/Refactor (API): Became more easier to create more request with Requester class inheritance in SpaceXAPI
